Fix lax.cond dtype mismatch in float32 mode
Cast sentinel value to params.dtype so both branches of lax.cond always
have matching output types, allowing float32 precision to work correctly.

+++ psytrax/models/ddm.py
@@ -81,7 +81,7 @@ def log_lik_trial(params, dat_trial):
     return lax.cond(
         valid,
         lambda _: _log_lik_valid(params, dat_trial),
-        lambda _: jnp.array(_INVALID_LOG_LIK),
+        lambda _: jnp.array(_INVALID_LOG_LIK, dtype=params.dtype),
         operand=None,
     )
